refactor: refactor utils (DD-63)
