+++ src/app/shared/utils/convertDateToISODateWithoutTime.ts
@@ -1,0 +1,7 @@
+export function convertDateToISODateWithoutTime(date: Date) {
+  const year = date.getUTCFullYear();
+  const month = date.getUTCMonth();
+  const day = date.getUTCDate();
+
+  return new Date(Date.UTC(year, month, day)).toISOString();
+}
