--- src/app/shared/utils/format-date.ts
@@ -1,6 +1,0 @@
-export function formatTime(dateStr: string): string {
-  const date = new Date(dateStr);
-  const hours = date.getUTCHours().toString().padStart(2, '0');
-  const minutes = date.getUTCMinutes().toString().padStart(2, '0');
-  return `${hours}:${minutes}`;
-}
